Fix bool not callable error and RuntimeWarning in module execution
This commit addresses two issues:

1. Fixes 'bool' object is not callable error in aider_ai_code.py by:
   - Creating SilentInputOutput subclass to properly override tool_error method
   - Adding top-level exception handling for this specific error

2. Fixes RuntimeWarning about __main__ execution by:
   - Moving CLI logic to a dedicated cli.py module
   - Making __main__.py a minimal entry point to reduce circular imports
   - Ensuring main function remains exposed in __init__.py for entry points
   - Updating tests to target cli.py instead of __main__.py
