+++ tests/test_main.py
@@ -7,7 +7,8 @@ import pytest
 from _pytest.tmpdir import TempPathFactory
 
 # Use absolute imports from the package root
-from aider_mcp_server.__main__ import main
+# Import main from cli instead of __main__ to prevent RuntimeWarning
+from aider_mcp_server.cli import main
 from aider_mcp_server.atoms.atoms_utils import (
     DEFAULT_EDITOR_MODEL,
 )
@@ -17,21 +18,21 @@ from aider_mcp_server.atoms.logging import Logger  # Import Logger for spec
 @pytest.fixture
 def mock_serve() -> Generator[AsyncMock, None, None]:
     """Mock the serve function (for stdio mode)."""
-    with patch("aider_mcp_server.__main__.serve", new_callable=AsyncMock) as mock:
+    with patch("aider_mcp_server.cli.serve", new_callable=AsyncMock) as mock:
         yield mock
 
 
 @pytest.fixture
 def mock_serve_sse() -> Generator[AsyncMock, None, None]:
     """Mock the serve_sse function."""
-    with patch("aider_mcp_server.__main__.serve_sse", new_callable=AsyncMock) as mock:
+    with patch("aider_mcp_server.cli.serve_sse", new_callable=AsyncMock) as mock:
         yield mock
 
 
 @pytest.fixture
 def mock_serve_multi() -> Generator[AsyncMock, None, None]:
     """Mock the serve_multi_transport function."""
-    with patch("aider_mcp_server.__main__.serve_multi_transport", new_callable=AsyncMock) as mock:
+    with patch("aider_mcp_server.cli.serve_multi_transport", new_callable=AsyncMock) as mock:
         yield mock
 
 
@@ -41,26 +42,26 @@ def mock_serve_multi() -> Generator[AsyncMock, None, None]:
 @pytest.fixture
 def mock_is_git_repository() -> Generator[MagicMock, None, None]:
     """Mock the is_git_repository check to always return True."""
-    # Patch where it's used in __main__
-    with patch("aider_mcp_server.__main__.is_git_repository", return_value=(True, None)) as mock:
+    # Patch where it's used in cli
+    with patch("aider_mcp_server.cli.is_git_repository", return_value=(True, None)) as mock:
         yield mock
 
 
 @pytest.fixture
 def mock_os_makedirs() -> Generator[MagicMock, None, None]:
     """Mock os.makedirs to avoid filesystem side effects."""
-    with patch("aider_mcp_server.__main__.os.makedirs") as mock:  # Patch where it's used
+    with patch("aider_mcp_server.cli.os.makedirs") as mock:  # Patch where it's used
         yield mock
 
 
 @pytest.fixture
 def mock_get_logger() -> Generator[MagicMock, None, None]:
     """Mock get_logger to return a mock logger with mocked methods."""
-    with patch("aider_mcp_server.__main__.get_logger", autospec=True) as mock_get_logger_factory:
+    with patch("aider_mcp_server.cli.get_logger", autospec=True) as mock_get_logger_factory:
         # Create a mock logger instance that conforms to the Logger protocol/class
         mock_logger_instance = MagicMock(spec=Logger)
-        # Ensure all methods used in __main__ are mocked
-        # Ensure all methods used in __main__ are mocked and will be tracked
+        # Ensure all methods used in cli are mocked
+        # Ensure all methods used in cli are mocked and will be tracked
         for method in ["debug", "info", "warning", "error", "critical", "exception"]:
             setattr(mock_logger_instance, method, MagicMock())
         # Set the return value of the factory mock
@@ -114,9 +115,9 @@ def run_main_with_args(
     # Mock signal handling setup
     mock_loop = MagicMock()
     mock_loop.add_signal_handler = MagicMock()
-    # Patch where get_event_loop is called (__main__)
-    with patch("aider_mcp_server.__main__.asyncio.get_event_loop", return_value=mock_loop):
-        # Patch Path.resolve and Path.is_dir as __main__ uses them
+    # Patch where get_event_loop is called (cli)
+    with patch("aider_mcp_server.cli.asyncio.get_event_loop", return_value=mock_loop):
+        # Patch Path.resolve and Path.is_dir as cli uses them
         with (
             patch.object(Path, "resolve") as mock_resolve,
             patch.object(Path, "is_dir") as mock_is_dir,
@@ -192,10 +193,10 @@ def test_invalid_cwd_not_git_repo(
     with patch.object(Path, "resolve", return_value=mock_resolved_path):
         # Mock is_git_repository to return False
         with patch(
-            "aider_mcp_server.__main__.is_git_repository",
+            "aider_mcp_server.cli.is_git_repository",
             return_value=(False, git_error_msg),
         ) as mock_git_check:
-            with patch("aider_mcp_server.__main__.get_logger") as mock_get_logger:
+            with patch("aider_mcp_server.cli.get_logger") as mock_get_logger:
                 mock_logger_instance = mock_get_logger.return_value
                 mock_logger_instance.critical = MagicMock()
